fix(PopOver): Fix PopOver freeze
In previous versions of react-popper, changes in the modifiers prop
were not detected & did not cause a rerender of the pop over.
This was changed in version 1.3.5, which unearthed a bug in our implementation
of it: When a modifier caused state to be set, this would cause a new modifier
object to be created, which would re-set the state,
and so forth, causing an infinite loop of updates.
This fix wraps the modifier update in a useMemo call, which prevents
this issue.